+++ ent/characterspellcasting/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.CharacterSpellcasting {
+func ID(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.CharacterSpellcasting {
+func IDEQ(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.CharacterSpellcasting {
+func IDNEQ(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.CharacterSpellcasting {
+func IDIn(ids ...int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.CharacterSpellcasting {
+func IDNotIn(ids ...int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.CharacterSpellcasting {
+func IDGT(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.CharacterSpellcasting {
+func IDGTE(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.CharacterSpellcasting {
+func IDLT(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.CharacterSpellcasting {
+func IDLTE(id int64) predicate.CharacterSpellcasting {
 	return predicate.CharacterSpellcasting(sql.FieldLTE(FieldID, id))
 }
